Refactored content signature
