+++ docs/docs/api/endpoints/messages-send.mdx
@@ -6,6 +6,17 @@ the payload will look differently for each source. Refer to each [source's docum
 to see learn how to send text, media, and many more message types.
 
 
+**Sample request**
+
+```json5
+{
+  "conversation_id": "a688d36c-a85e-44af-bc02-4248c2c97622",
+  "message": {
+    // Source specific body
+  }
+}
+```
+
 **Sample response**
 
 ```json5
