+++ docs/sidebars.js
@@ -61,6 +61,7 @@ module.exports = {
         },
         'api/websocket',
         'api/webhook',
+        'api/source',
       ],
     },
     {
